Find latest version of a file

If some file names include dates, this function will find the latest version of a given file. This can be handy e.g. for finding the latest data update automatically. find_latest scans recursively a directory, looking for file names matching a given pattern (or regular expression) and including dates. Dates are extracted using extract_date, using a user-provided parser (see details). The full path to the most recent file will be returned.

find_latest(
  pattern,
  where = getwd(),
  date_parser = lubridate::ymd,
  quiet = FALSE,
  ...
)

Arguments

pattern
a character string indicating the name of the file to look for, or regular expression to be matched against file names

where
the directory in which to look for the file; defaults to the current directory

date_parser
a function to extract dates from character strings; we recommend using lubridate's function there, e.g. ymd (any date format with year, month, day) or dmy (any format using day, month, year); defaults to lubridate::ymd

quiet
a logical indicating if messages should be displayed (TRUE, default), or not (FALSE)

...
further arguments passed to list.files

Details

The date parser used by default is lubridate::ymd, so that any date specified as year/month/day (regardless of the separator used) should work. One exception is that only dates provided as numbers can be used e.g. "1982.02.04" and "82/2/4" are okay but not "1982 Feb 04". For dates provided in a different order, check other parsers implemented in lubridate, e.g. lubridate::dmy for day-month-year formats, or lubridate::ydm for year-day-month. Note that parsers for date-time objects can also be used, but in any case a Date object will be returned.

See also

Author

Thibaut Jombart

Examples

 

if (require(reportfactory)) {

## create random factory and toy files
odir <- getwd()
f <- new_factory(path = tempdir())

file.create(file.path("data", "linelist_2020-10-01.xlsx"))
file.create(file.path("data", "linelist_2020-10-12.csv"))
file.create(file.path("data", "linelist.xlsx"))
file.create(file.path("data", "contacts.xlsx"))
file.create(file.path("data", "death_linelist_2020-10-13.xlsx"))

## find the latest data with 'linelist' in the name; note that this
## matches both 'linelist' and 'death_linelist' files
find_latest("linelist")

## same, but this time files starting with 'linelist', i.e. excluding
## 'death_linelist'
find_latest("^linelist")

## this returns NULL
find_latest("foobar")

## cleanup
unlink(f, recursive = TRUE)
setwd(odir)
}

#> Loading required package: reportfactory

#> lubridate::ymd could not find dates in files matching linelist

#> Found file: 'death_linelist_2020-10-13.xlsx'

#> lubridate::ymd could not find dates in files matching ^linelist

#> Found file: 'linelist_2020-10-12.csv'

#> No file matching pattern 'foobar' found in '/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpvPrOvJ/new_factory'